[WIP] tee moved to modules
