+++ src/proto/api.ts
@@ -462,11 +462,7 @@ export interface MessageReveal {
     | MessageReveal_MessageRevealDirect
     | undefined;
   /** partially or fully reveal the block via a zk proof */
-  zkReveal?:
-    | MessageReveal_MessageRevealZk
-    | undefined;
-  /** Reveals of TEE services */
-  teeStreamReveal?: MessageReveal_MessageRevealTeeStream | undefined;
+  zkReveal?: MessageReveal_MessageRevealZk | undefined;
 }
 
 export interface MessageReveal_MessageRevealDirect {
@@ -484,15 +480,6 @@ export interface MessageReveal_MessageRevealDirect {
 export interface MessageReveal_MessageRevealZk {
   proofs: MessageReveal_ZKProof[];
   toprfs: MessageReveal_TOPRFProof[];
-}
-
-export interface MessageReveal_MessageRevealTeeStream {
-  /** The revealed (redacted) data from TEE */
-  revealedData: Uint8Array;
-  /** TEE signature over this data */
-  teeSignature: Uint8Array;
-  /** TEE public key for verification */
-  teePublicKey: Uint8Array;
 }
 
 export interface MessageReveal_ZKProof {
@@ -552,39 +539,6 @@ export interface ClaimRequestData {
    */
   timestampS: number;
   context: string;
-}
-
-export interface ClaimTeeBundleRequest {
-  verificationBundle: Uint8Array;
-  data: ClaimRequestData | undefined;
-  signatures: ClaimTeeBundleRequest_Signatures | undefined;
-}
-
-export interface ClaimTeeBundleRequest_Signatures {
-  requestSignature: Uint8Array;
-}
-
-export interface ClaimTeeBundleResponse {
-  /** The original request that was made */
-  request: ClaimTeeBundleRequest | undefined;
-  claim?: ProviderClaimData | undefined;
-  error?: ErrorData | undefined;
-  signatures: ClaimTeeBundleResponse_Signatures | undefined;
-}
-
-export interface ClaimTeeBundleResponse_Signatures {
-  /** Address of the attestor that has signed the claim */
-  attestorAddress: string;
-  /**
-   * Signature of `stringifyProviderClaimData(claim)`,
-   * if the claim was successful
-   */
-  claimSignature: Uint8Array;
-  /**
-   * Signature of the complete ClaimTeeBundleResponse
-   * structure with empty "signatures" field
-   */
-  resultSignature: Uint8Array;
 }
 
 export interface ClaimTunnelRequest {
@@ -755,6 +709,39 @@ export interface TOPRFResponse {
   evaluated: Uint8Array;
   c: Uint8Array;
   r: Uint8Array;
+}
+
+export interface ClaimTeeBundleRequest {
+  verificationBundle: Uint8Array;
+  data: ClaimRequestData | undefined;
+  signatures: ClaimTeeBundleRequest_Signatures | undefined;
+}
+
+export interface ClaimTeeBundleRequest_Signatures {
+  requestSignature: Uint8Array;
+}
+
+export interface ClaimTeeBundleResponse {
+  /** The original request that was made */
+  request: ClaimTeeBundleRequest | undefined;
+  claim?: ProviderClaimData | undefined;
+  error?: ErrorData | undefined;
+  signatures: ClaimTeeBundleResponse_Signatures | undefined;
+}
+
+export interface ClaimTeeBundleResponse_Signatures {
+  /** Address of the attestor that has signed the claim */
+  attestorAddress: string;
+  /**
+   * Signature of `stringifyProviderClaimData(claim)`,
+   * if the claim was successful
+   */
+  claimSignature: Uint8Array;
+  /**
+   * Signature of the complete ClaimTeeBundleResponse
+   * structure with empty "signatures" field
+   */
+  resultSignature: Uint8Array;
 }
 
 export interface RPCMessage {
@@ -1732,7 +1719,7 @@ export const TunnelDisconnectEvent: MessageFns<TunnelDisconnectEvent> = {
 };
 
 function createBaseMessageReveal(): MessageReveal {
-  return { directReveal: undefined, zkReveal: undefined, teeStreamReveal: undefined };
+  return { directReveal: undefined, zkReveal: undefined };
 }
 
 export const MessageReveal: MessageFns<MessageReveal> = {
@@ -1742,9 +1729,6 @@ export const MessageReveal: MessageFns<MessageReveal> = {
     }
     if (message.zkReveal !== undefined) {
       MessageReveal_MessageRevealZk.encode(message.zkReveal, writer.uint32(18).fork()).join();
-    }
-    if (message.teeStreamReveal !== undefined) {
-      MessageReveal_MessageRevealTeeStream.encode(message.teeStreamReveal, writer.uint32(26).fork()).join();
     }
     return writer;
   },
@@ -1772,14 +1756,6 @@ export const MessageReveal: MessageFns<MessageReveal> = {
           message.zkReveal = MessageReveal_MessageRevealZk.decode(reader, reader.uint32());
           continue;
         }
-        case 3: {
-          if (tag !== 26) {
-            break;
-          }
-
-          message.teeStreamReveal = MessageReveal_MessageRevealTeeStream.decode(reader, reader.uint32());
-          continue;
-        }
       }
       if ((tag & 7) === 4 || tag === 0) {
         break;
@@ -1795,9 +1771,6 @@ export const MessageReveal: MessageFns<MessageReveal> = {
         ? MessageReveal_MessageRevealDirect.fromJSON(object.directReveal)
         : undefined,
       zkReveal: isSet(object.zkReveal) ? MessageReveal_MessageRevealZk.fromJSON(object.zkReveal) : undefined,
-      teeStreamReveal: isSet(object.teeStreamReveal)
-        ? MessageReveal_MessageRevealTeeStream.fromJSON(object.teeStreamReveal)
-        : undefined,
     };
   },
 
@@ -1808,9 +1781,6 @@ export const MessageReveal: MessageFns<MessageReveal> = {
     }
     if (message.zkReveal !== undefined) {
       obj.zkReveal = MessageReveal_MessageRevealZk.toJSON(message.zkReveal);
-    }
-    if (message.teeStreamReveal !== undefined) {
-      obj.teeStreamReveal = MessageReveal_MessageRevealTeeStream.toJSON(message.teeStreamReveal);
     }
     return obj;
   },
@@ -1825,9 +1795,6 @@ export const MessageReveal: MessageFns<MessageReveal> = {
       : undefined;
     message.zkReveal = (object.zkReveal !== undefined && object.zkReveal !== null)
       ? MessageReveal_MessageRevealZk.fromPartial(object.zkReveal)
-      : undefined;
-    message.teeStreamReveal = (object.teeStreamReveal !== undefined && object.teeStreamReveal !== null)
-      ? MessageReveal_MessageRevealTeeStream.fromPartial(object.teeStreamReveal)
       : undefined;
     return message;
   },
@@ -2001,98 +1968,6 @@ export const MessageReveal_MessageRevealZk: MessageFns<MessageReveal_MessageReve
     const message = createBaseMessageReveal_MessageRevealZk();
     message.proofs = object.proofs?.map((e) => MessageReveal_ZKProof.fromPartial(e)) || [];
     message.toprfs = object.toprfs?.map((e) => MessageReveal_TOPRFProof.fromPartial(e)) || [];
-    return message;
-  },
-};
-
-function createBaseMessageReveal_MessageRevealTeeStream(): MessageReveal_MessageRevealTeeStream {
-  return { revealedData: new Uint8Array(0), teeSignature: new Uint8Array(0), teePublicKey: new Uint8Array(0) };
-}
-
-export const MessageReveal_MessageRevealTeeStream: MessageFns<MessageReveal_MessageRevealTeeStream> = {
-  encode(message: MessageReveal_MessageRevealTeeStream, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.revealedData.length !== 0) {
-      writer.uint32(10).bytes(message.revealedData);
-    }
-    if (message.teeSignature.length !== 0) {
-      writer.uint32(18).bytes(message.teeSignature);
-    }
-    if (message.teePublicKey.length !== 0) {
-      writer.uint32(26).bytes(message.teePublicKey);
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): MessageReveal_MessageRevealTeeStream {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseMessageReveal_MessageRevealTeeStream();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.revealedData = reader.bytes();
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.teeSignature = reader.bytes();
-          continue;
-        }
-        case 3: {
-          if (tag !== 26) {
-            break;
-          }
-
-          message.teePublicKey = reader.bytes();
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): MessageReveal_MessageRevealTeeStream {
-    return {
-      revealedData: isSet(object.revealedData) ? bytesFromBase64(object.revealedData) : new Uint8Array(0),
-      teeSignature: isSet(object.teeSignature) ? bytesFromBase64(object.teeSignature) : new Uint8Array(0),
-      teePublicKey: isSet(object.teePublicKey) ? bytesFromBase64(object.teePublicKey) : new Uint8Array(0),
-    };
-  },
-
-  toJSON(message: MessageReveal_MessageRevealTeeStream): unknown {
-    const obj: any = {};
-    if (message.revealedData.length !== 0) {
-      obj.revealedData = base64FromBytes(message.revealedData);
-    }
-    if (message.teeSignature.length !== 0) {
-      obj.teeSignature = base64FromBytes(message.teeSignature);
-    }
-    if (message.teePublicKey.length !== 0) {
-      obj.teePublicKey = base64FromBytes(message.teePublicKey);
-    }
-    return obj;
-  },
-
-  create(base?: DeepPartial<MessageReveal_MessageRevealTeeStream>): MessageReveal_MessageRevealTeeStream {
-    return MessageReveal_MessageRevealTeeStream.fromPartial(base ?? {});
-  },
-  fromPartial(object: DeepPartial<MessageReveal_MessageRevealTeeStream>): MessageReveal_MessageRevealTeeStream {
-    const message = createBaseMessageReveal_MessageRevealTeeStream();
-    message.revealedData = object.revealedData ?? new Uint8Array(0);
-    message.teeSignature = object.teeSignature ?? new Uint8Array(0);
-    message.teePublicKey = object.teePublicKey ?? new Uint8Array(0);
     return message;
   },
 };
@@ -2600,372 +2475,6 @@ export const ClaimRequestData: MessageFns<ClaimRequestData> = {
     message.owner = object.owner ?? "";
     message.timestampS = object.timestampS ?? 0;
     message.context = object.context ?? "";
-    return message;
-  },
-};
-
-function createBaseClaimTeeBundleRequest(): ClaimTeeBundleRequest {
-  return { verificationBundle: new Uint8Array(0), data: undefined, signatures: undefined };
-}
-
-export const ClaimTeeBundleRequest: MessageFns<ClaimTeeBundleRequest> = {
-  encode(message: ClaimTeeBundleRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.verificationBundle.length !== 0) {
-      writer.uint32(10).bytes(message.verificationBundle);
-    }
-    if (message.data !== undefined) {
-      ClaimRequestData.encode(message.data, writer.uint32(18).fork()).join();
-    }
-    if (message.signatures !== undefined) {
-      ClaimTeeBundleRequest_Signatures.encode(message.signatures, writer.uint32(26).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleRequest {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseClaimTeeBundleRequest();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.verificationBundle = reader.bytes();
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.data = ClaimRequestData.decode(reader, reader.uint32());
-          continue;
-        }
-        case 3: {
-          if (tag !== 26) {
-            break;
-          }
-
-          message.signatures = ClaimTeeBundleRequest_Signatures.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): ClaimTeeBundleRequest {
-    return {
-      verificationBundle: isSet(object.verificationBundle)
-        ? bytesFromBase64(object.verificationBundle)
-        : new Uint8Array(0),
-      data: isSet(object.data) ? ClaimRequestData.fromJSON(object.data) : undefined,
-      signatures: isSet(object.signatures) ? ClaimTeeBundleRequest_Signatures.fromJSON(object.signatures) : undefined,
-    };
-  },
-
-  toJSON(message: ClaimTeeBundleRequest): unknown {
-    const obj: any = {};
-    if (message.verificationBundle.length !== 0) {
-      obj.verificationBundle = base64FromBytes(message.verificationBundle);
-    }
-    if (message.data !== undefined) {
-      obj.data = ClaimRequestData.toJSON(message.data);
-    }
-    if (message.signatures !== undefined) {
-      obj.signatures = ClaimTeeBundleRequest_Signatures.toJSON(message.signatures);
-    }
-    return obj;
-  },
-
-  create(base?: DeepPartial<ClaimTeeBundleRequest>): ClaimTeeBundleRequest {
-    return ClaimTeeBundleRequest.fromPartial(base ?? {});
-  },
-  fromPartial(object: DeepPartial<ClaimTeeBundleRequest>): ClaimTeeBundleRequest {
-    const message = createBaseClaimTeeBundleRequest();
-    message.verificationBundle = object.verificationBundle ?? new Uint8Array(0);
-    message.data = (object.data !== undefined && object.data !== null)
-      ? ClaimRequestData.fromPartial(object.data)
-      : undefined;
-    message.signatures = (object.signatures !== undefined && object.signatures !== null)
-      ? ClaimTeeBundleRequest_Signatures.fromPartial(object.signatures)
-      : undefined;
-    return message;
-  },
-};
-
-function createBaseClaimTeeBundleRequest_Signatures(): ClaimTeeBundleRequest_Signatures {
-  return { requestSignature: new Uint8Array(0) };
-}
-
-export const ClaimTeeBundleRequest_Signatures: MessageFns<ClaimTeeBundleRequest_Signatures> = {
-  encode(message: ClaimTeeBundleRequest_Signatures, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.requestSignature.length !== 0) {
-      writer.uint32(10).bytes(message.requestSignature);
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleRequest_Signatures {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseClaimTeeBundleRequest_Signatures();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.requestSignature = reader.bytes();
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): ClaimTeeBundleRequest_Signatures {
-    return {
-      requestSignature: isSet(object.requestSignature) ? bytesFromBase64(object.requestSignature) : new Uint8Array(0),
-    };
-  },
-
-  toJSON(message: ClaimTeeBundleRequest_Signatures): unknown {
-    const obj: any = {};
-    if (message.requestSignature.length !== 0) {
-      obj.requestSignature = base64FromBytes(message.requestSignature);
-    }
-    return obj;
-  },
-
-  create(base?: DeepPartial<ClaimTeeBundleRequest_Signatures>): ClaimTeeBundleRequest_Signatures {
-    return ClaimTeeBundleRequest_Signatures.fromPartial(base ?? {});
-  },
-  fromPartial(object: DeepPartial<ClaimTeeBundleRequest_Signatures>): ClaimTeeBundleRequest_Signatures {
-    const message = createBaseClaimTeeBundleRequest_Signatures();
-    message.requestSignature = object.requestSignature ?? new Uint8Array(0);
-    return message;
-  },
-};
-
-function createBaseClaimTeeBundleResponse(): ClaimTeeBundleResponse {
-  return { request: undefined, claim: undefined, error: undefined, signatures: undefined };
-}
-
-export const ClaimTeeBundleResponse: MessageFns<ClaimTeeBundleResponse> = {
-  encode(message: ClaimTeeBundleResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.request !== undefined) {
-      ClaimTeeBundleRequest.encode(message.request, writer.uint32(10).fork()).join();
-    }
-    if (message.claim !== undefined) {
-      ProviderClaimData.encode(message.claim, writer.uint32(18).fork()).join();
-    }
-    if (message.error !== undefined) {
-      ErrorData.encode(message.error, writer.uint32(26).fork()).join();
-    }
-    if (message.signatures !== undefined) {
-      ClaimTeeBundleResponse_Signatures.encode(message.signatures, writer.uint32(34).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleResponse {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseClaimTeeBundleResponse();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.request = ClaimTeeBundleRequest.decode(reader, reader.uint32());
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.claim = ProviderClaimData.decode(reader, reader.uint32());
-          continue;
-        }
-        case 3: {
-          if (tag !== 26) {
-            break;
-          }
-
-          message.error = ErrorData.decode(reader, reader.uint32());
-          continue;
-        }
-        case 4: {
-          if (tag !== 34) {
-            break;
-          }
-
-          message.signatures = ClaimTeeBundleResponse_Signatures.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): ClaimTeeBundleResponse {
-    return {
-      request: isSet(object.request) ? ClaimTeeBundleRequest.fromJSON(object.request) : undefined,
-      claim: isSet(object.claim) ? ProviderClaimData.fromJSON(object.claim) : undefined,
-      error: isSet(object.error) ? ErrorData.fromJSON(object.error) : undefined,
-      signatures: isSet(object.signatures) ? ClaimTeeBundleResponse_Signatures.fromJSON(object.signatures) : undefined,
-    };
-  },
-
-  toJSON(message: ClaimTeeBundleResponse): unknown {
-    const obj: any = {};
-    if (message.request !== undefined) {
-      obj.request = ClaimTeeBundleRequest.toJSON(message.request);
-    }
-    if (message.claim !== undefined) {
-      obj.claim = ProviderClaimData.toJSON(message.claim);
-    }
-    if (message.error !== undefined) {
-      obj.error = ErrorData.toJSON(message.error);
-    }
-    if (message.signatures !== undefined) {
-      obj.signatures = ClaimTeeBundleResponse_Signatures.toJSON(message.signatures);
-    }
-    return obj;
-  },
-
-  create(base?: DeepPartial<ClaimTeeBundleResponse>): ClaimTeeBundleResponse {
-    return ClaimTeeBundleResponse.fromPartial(base ?? {});
-  },
-  fromPartial(object: DeepPartial<ClaimTeeBundleResponse>): ClaimTeeBundleResponse {
-    const message = createBaseClaimTeeBundleResponse();
-    message.request = (object.request !== undefined && object.request !== null)
-      ? ClaimTeeBundleRequest.fromPartial(object.request)
-      : undefined;
-    message.claim = (object.claim !== undefined && object.claim !== null)
-      ? ProviderClaimData.fromPartial(object.claim)
-      : undefined;
-    message.error = (object.error !== undefined && object.error !== null)
-      ? ErrorData.fromPartial(object.error)
-      : undefined;
-    message.signatures = (object.signatures !== undefined && object.signatures !== null)
-      ? ClaimTeeBundleResponse_Signatures.fromPartial(object.signatures)
-      : undefined;
-    return message;
-  },
-};
-
-function createBaseClaimTeeBundleResponse_Signatures(): ClaimTeeBundleResponse_Signatures {
-  return { attestorAddress: "", claimSignature: new Uint8Array(0), resultSignature: new Uint8Array(0) };
-}
-
-export const ClaimTeeBundleResponse_Signatures: MessageFns<ClaimTeeBundleResponse_Signatures> = {
-  encode(message: ClaimTeeBundleResponse_Signatures, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.attestorAddress !== "") {
-      writer.uint32(10).string(message.attestorAddress);
-    }
-    if (message.claimSignature.length !== 0) {
-      writer.uint32(18).bytes(message.claimSignature);
-    }
-    if (message.resultSignature.length !== 0) {
-      writer.uint32(26).bytes(message.resultSignature);
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleResponse_Signatures {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    let end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseClaimTeeBundleResponse_Signatures();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.attestorAddress = reader.string();
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.claimSignature = reader.bytes();
-          continue;
-        }
-        case 3: {
-          if (tag !== 26) {
-            break;
-          }
-
-          message.resultSignature = reader.bytes();
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): ClaimTeeBundleResponse_Signatures {
-    return {
-      attestorAddress: isSet(object.attestorAddress) ? globalThis.String(object.attestorAddress) : "",
-      claimSignature: isSet(object.claimSignature) ? bytesFromBase64(object.claimSignature) : new Uint8Array(0),
-      resultSignature: isSet(object.resultSignature) ? bytesFromBase64(object.resultSignature) : new Uint8Array(0),
-    };
-  },
-
-  toJSON(message: ClaimTeeBundleResponse_Signatures): unknown {
-    const obj: any = {};
-    if (message.attestorAddress !== "") {
-      obj.attestorAddress = message.attestorAddress;
-    }
-    if (message.claimSignature.length !== 0) {
-      obj.claimSignature = base64FromBytes(message.claimSignature);
-    }
-    if (message.resultSignature.length !== 0) {
-      obj.resultSignature = base64FromBytes(message.resultSignature);
-    }
-    return obj;
-  },
-
-  create(base?: DeepPartial<ClaimTeeBundleResponse_Signatures>): ClaimTeeBundleResponse_Signatures {
-    return ClaimTeeBundleResponse_Signatures.fromPartial(base ?? {});
-  },
-  fromPartial(object: DeepPartial<ClaimTeeBundleResponse_Signatures>): ClaimTeeBundleResponse_Signatures {
-    const message = createBaseClaimTeeBundleResponse_Signatures();
-    message.attestorAddress = object.attestorAddress ?? "";
-    message.claimSignature = object.claimSignature ?? new Uint8Array(0);
-    message.resultSignature = object.resultSignature ?? new Uint8Array(0);
     return message;
   },
 };
@@ -4537,6 +4046,372 @@ export const TOPRFResponse: MessageFns<TOPRFResponse> = {
     message.evaluated = object.evaluated ?? new Uint8Array(0);
     message.c = object.c ?? new Uint8Array(0);
     message.r = object.r ?? new Uint8Array(0);
+    return message;
+  },
+};
+
+function createBaseClaimTeeBundleRequest(): ClaimTeeBundleRequest {
+  return { verificationBundle: new Uint8Array(0), data: undefined, signatures: undefined };
+}
+
+export const ClaimTeeBundleRequest: MessageFns<ClaimTeeBundleRequest> = {
+  encode(message: ClaimTeeBundleRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
+    if (message.verificationBundle.length !== 0) {
+      writer.uint32(10).bytes(message.verificationBundle);
+    }
+    if (message.data !== undefined) {
+      ClaimRequestData.encode(message.data, writer.uint32(18).fork()).join();
+    }
+    if (message.signatures !== undefined) {
+      ClaimTeeBundleRequest_Signatures.encode(message.signatures, writer.uint32(26).fork()).join();
+    }
+    return writer;
+  },
+
+  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleRequest {
+    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseClaimTeeBundleRequest();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1: {
+          if (tag !== 10) {
+            break;
+          }
+
+          message.verificationBundle = reader.bytes();
+          continue;
+        }
+        case 2: {
+          if (tag !== 18) {
+            break;
+          }
+
+          message.data = ClaimRequestData.decode(reader, reader.uint32());
+          continue;
+        }
+        case 3: {
+          if (tag !== 26) {
+            break;
+          }
+
+          message.signatures = ClaimTeeBundleRequest_Signatures.decode(reader, reader.uint32());
+          continue;
+        }
+      }
+      if ((tag & 7) === 4 || tag === 0) {
+        break;
+      }
+      reader.skip(tag & 7);
+    }
+    return message;
+  },
+
+  fromJSON(object: any): ClaimTeeBundleRequest {
+    return {
+      verificationBundle: isSet(object.verificationBundle)
+        ? bytesFromBase64(object.verificationBundle)
+        : new Uint8Array(0),
+      data: isSet(object.data) ? ClaimRequestData.fromJSON(object.data) : undefined,
+      signatures: isSet(object.signatures) ? ClaimTeeBundleRequest_Signatures.fromJSON(object.signatures) : undefined,
+    };
+  },
+
+  toJSON(message: ClaimTeeBundleRequest): unknown {
+    const obj: any = {};
+    if (message.verificationBundle.length !== 0) {
+      obj.verificationBundle = base64FromBytes(message.verificationBundle);
+    }
+    if (message.data !== undefined) {
+      obj.data = ClaimRequestData.toJSON(message.data);
+    }
+    if (message.signatures !== undefined) {
+      obj.signatures = ClaimTeeBundleRequest_Signatures.toJSON(message.signatures);
+    }
+    return obj;
+  },
+
+  create(base?: DeepPartial<ClaimTeeBundleRequest>): ClaimTeeBundleRequest {
+    return ClaimTeeBundleRequest.fromPartial(base ?? {});
+  },
+  fromPartial(object: DeepPartial<ClaimTeeBundleRequest>): ClaimTeeBundleRequest {
+    const message = createBaseClaimTeeBundleRequest();
+    message.verificationBundle = object.verificationBundle ?? new Uint8Array(0);
+    message.data = (object.data !== undefined && object.data !== null)
+      ? ClaimRequestData.fromPartial(object.data)
+      : undefined;
+    message.signatures = (object.signatures !== undefined && object.signatures !== null)
+      ? ClaimTeeBundleRequest_Signatures.fromPartial(object.signatures)
+      : undefined;
+    return message;
+  },
+};
+
+function createBaseClaimTeeBundleRequest_Signatures(): ClaimTeeBundleRequest_Signatures {
+  return { requestSignature: new Uint8Array(0) };
+}
+
+export const ClaimTeeBundleRequest_Signatures: MessageFns<ClaimTeeBundleRequest_Signatures> = {
+  encode(message: ClaimTeeBundleRequest_Signatures, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
+    if (message.requestSignature.length !== 0) {
+      writer.uint32(10).bytes(message.requestSignature);
+    }
+    return writer;
+  },
+
+  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleRequest_Signatures {
+    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseClaimTeeBundleRequest_Signatures();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1: {
+          if (tag !== 10) {
+            break;
+          }
+
+          message.requestSignature = reader.bytes();
+          continue;
+        }
+      }
+      if ((tag & 7) === 4 || tag === 0) {
+        break;
+      }
+      reader.skip(tag & 7);
+    }
+    return message;
+  },
+
+  fromJSON(object: any): ClaimTeeBundleRequest_Signatures {
+    return {
+      requestSignature: isSet(object.requestSignature) ? bytesFromBase64(object.requestSignature) : new Uint8Array(0),
+    };
+  },
+
+  toJSON(message: ClaimTeeBundleRequest_Signatures): unknown {
+    const obj: any = {};
+    if (message.requestSignature.length !== 0) {
+      obj.requestSignature = base64FromBytes(message.requestSignature);
+    }
+    return obj;
+  },
+
+  create(base?: DeepPartial<ClaimTeeBundleRequest_Signatures>): ClaimTeeBundleRequest_Signatures {
+    return ClaimTeeBundleRequest_Signatures.fromPartial(base ?? {});
+  },
+  fromPartial(object: DeepPartial<ClaimTeeBundleRequest_Signatures>): ClaimTeeBundleRequest_Signatures {
+    const message = createBaseClaimTeeBundleRequest_Signatures();
+    message.requestSignature = object.requestSignature ?? new Uint8Array(0);
+    return message;
+  },
+};
+
+function createBaseClaimTeeBundleResponse(): ClaimTeeBundleResponse {
+  return { request: undefined, claim: undefined, error: undefined, signatures: undefined };
+}
+
+export const ClaimTeeBundleResponse: MessageFns<ClaimTeeBundleResponse> = {
+  encode(message: ClaimTeeBundleResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
+    if (message.request !== undefined) {
+      ClaimTeeBundleRequest.encode(message.request, writer.uint32(10).fork()).join();
+    }
+    if (message.claim !== undefined) {
+      ProviderClaimData.encode(message.claim, writer.uint32(18).fork()).join();
+    }
+    if (message.error !== undefined) {
+      ErrorData.encode(message.error, writer.uint32(26).fork()).join();
+    }
+    if (message.signatures !== undefined) {
+      ClaimTeeBundleResponse_Signatures.encode(message.signatures, writer.uint32(34).fork()).join();
+    }
+    return writer;
+  },
+
+  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleResponse {
+    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseClaimTeeBundleResponse();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1: {
+          if (tag !== 10) {
+            break;
+          }
+
+          message.request = ClaimTeeBundleRequest.decode(reader, reader.uint32());
+          continue;
+        }
+        case 2: {
+          if (tag !== 18) {
+            break;
+          }
+
+          message.claim = ProviderClaimData.decode(reader, reader.uint32());
+          continue;
+        }
+        case 3: {
+          if (tag !== 26) {
+            break;
+          }
+
+          message.error = ErrorData.decode(reader, reader.uint32());
+          continue;
+        }
+        case 4: {
+          if (tag !== 34) {
+            break;
+          }
+
+          message.signatures = ClaimTeeBundleResponse_Signatures.decode(reader, reader.uint32());
+          continue;
+        }
+      }
+      if ((tag & 7) === 4 || tag === 0) {
+        break;
+      }
+      reader.skip(tag & 7);
+    }
+    return message;
+  },
+
+  fromJSON(object: any): ClaimTeeBundleResponse {
+    return {
+      request: isSet(object.request) ? ClaimTeeBundleRequest.fromJSON(object.request) : undefined,
+      claim: isSet(object.claim) ? ProviderClaimData.fromJSON(object.claim) : undefined,
+      error: isSet(object.error) ? ErrorData.fromJSON(object.error) : undefined,
+      signatures: isSet(object.signatures) ? ClaimTeeBundleResponse_Signatures.fromJSON(object.signatures) : undefined,
+    };
+  },
+
+  toJSON(message: ClaimTeeBundleResponse): unknown {
+    const obj: any = {};
+    if (message.request !== undefined) {
+      obj.request = ClaimTeeBundleRequest.toJSON(message.request);
+    }
+    if (message.claim !== undefined) {
+      obj.claim = ProviderClaimData.toJSON(message.claim);
+    }
+    if (message.error !== undefined) {
+      obj.error = ErrorData.toJSON(message.error);
+    }
+    if (message.signatures !== undefined) {
+      obj.signatures = ClaimTeeBundleResponse_Signatures.toJSON(message.signatures);
+    }
+    return obj;
+  },
+
+  create(base?: DeepPartial<ClaimTeeBundleResponse>): ClaimTeeBundleResponse {
+    return ClaimTeeBundleResponse.fromPartial(base ?? {});
+  },
+  fromPartial(object: DeepPartial<ClaimTeeBundleResponse>): ClaimTeeBundleResponse {
+    const message = createBaseClaimTeeBundleResponse();
+    message.request = (object.request !== undefined && object.request !== null)
+      ? ClaimTeeBundleRequest.fromPartial(object.request)
+      : undefined;
+    message.claim = (object.claim !== undefined && object.claim !== null)
+      ? ProviderClaimData.fromPartial(object.claim)
+      : undefined;
+    message.error = (object.error !== undefined && object.error !== null)
+      ? ErrorData.fromPartial(object.error)
+      : undefined;
+    message.signatures = (object.signatures !== undefined && object.signatures !== null)
+      ? ClaimTeeBundleResponse_Signatures.fromPartial(object.signatures)
+      : undefined;
+    return message;
+  },
+};
+
+function createBaseClaimTeeBundleResponse_Signatures(): ClaimTeeBundleResponse_Signatures {
+  return { attestorAddress: "", claimSignature: new Uint8Array(0), resultSignature: new Uint8Array(0) };
+}
+
+export const ClaimTeeBundleResponse_Signatures: MessageFns<ClaimTeeBundleResponse_Signatures> = {
+  encode(message: ClaimTeeBundleResponse_Signatures, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
+    if (message.attestorAddress !== "") {
+      writer.uint32(10).string(message.attestorAddress);
+    }
+    if (message.claimSignature.length !== 0) {
+      writer.uint32(18).bytes(message.claimSignature);
+    }
+    if (message.resultSignature.length !== 0) {
+      writer.uint32(26).bytes(message.resultSignature);
+    }
+    return writer;
+  },
+
+  decode(input: BinaryReader | Uint8Array, length?: number): ClaimTeeBundleResponse_Signatures {
+    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseClaimTeeBundleResponse_Signatures();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1: {
+          if (tag !== 10) {
+            break;
+          }
+
+          message.attestorAddress = reader.string();
+          continue;
+        }
+        case 2: {
+          if (tag !== 18) {
+            break;
+          }
+
+          message.claimSignature = reader.bytes();
+          continue;
+        }
+        case 3: {
+          if (tag !== 26) {
+            break;
+          }
+
+          message.resultSignature = reader.bytes();
+          continue;
+        }
+      }
+      if ((tag & 7) === 4 || tag === 0) {
+        break;
+      }
+      reader.skip(tag & 7);
+    }
+    return message;
+  },
+
+  fromJSON(object: any): ClaimTeeBundleResponse_Signatures {
+    return {
+      attestorAddress: isSet(object.attestorAddress) ? globalThis.String(object.attestorAddress) : "",
+      claimSignature: isSet(object.claimSignature) ? bytesFromBase64(object.claimSignature) : new Uint8Array(0),
+      resultSignature: isSet(object.resultSignature) ? bytesFromBase64(object.resultSignature) : new Uint8Array(0),
+    };
+  },
+
+  toJSON(message: ClaimTeeBundleResponse_Signatures): unknown {
+    const obj: any = {};
+    if (message.attestorAddress !== "") {
+      obj.attestorAddress = message.attestorAddress;
+    }
+    if (message.claimSignature.length !== 0) {
+      obj.claimSignature = base64FromBytes(message.claimSignature);
+    }
+    if (message.resultSignature.length !== 0) {
+      obj.resultSignature = base64FromBytes(message.resultSignature);
+    }
+    return obj;
+  },
+
+  create(base?: DeepPartial<ClaimTeeBundleResponse_Signatures>): ClaimTeeBundleResponse_Signatures {
+    return ClaimTeeBundleResponse_Signatures.fromPartial(base ?? {});
+  },
+  fromPartial(object: DeepPartial<ClaimTeeBundleResponse_Signatures>): ClaimTeeBundleResponse_Signatures {
+    const message = createBaseClaimTeeBundleResponse_Signatures();
+    message.attestorAddress = object.attestorAddress ?? "";
+    message.claimSignature = object.claimSignature ?? new Uint8Array(0);
+    message.resultSignature = object.resultSignature ?? new Uint8Array(0);
     return message;
   },
 };
