+++ src/server/handlers/index.ts
@@ -1,13 +1,13 @@
-import { claimTeeBundle } from 'src/server/handlers/claimTeeBundle'
-import { claimTunnel } from 'src/server/handlers/claimTunnel'
-import { completeClaimOnChain } from 'src/server/handlers/completeClaimOnChain'
-import { createClaimOnChain } from 'src/server/handlers/createClaimOnChain'
-import { createTaskOnMechain } from 'src/server/handlers/createTaskOnMechain'
-import { createTunnel } from 'src/server/handlers/createTunnel'
-import { disconnectTunnel } from 'src/server/handlers/disconnectTunnel'
-import { init } from 'src/server/handlers/init'
-import { toprf } from 'src/server/handlers/toprf'
-import { RPCHandler, RPCType } from 'src/types'
+import { claimTeeBundle } from '#src/server/handlers/claimTeeBundle.ts'
+import { claimTunnel } from '#src/server/handlers/claimTunnel.ts'
+import { completeClaimOnChain } from '#src/server/handlers/completeClaimOnChain.ts'
+import { createClaimOnChain } from '#src/server/handlers/createClaimOnChain.ts'
+import { createTaskOnMechain } from '#src/server/handlers/createTaskOnMechain.ts'
+import { createTunnel } from '#src/server/handlers/createTunnel.ts'
+import { disconnectTunnel } from '#src/server/handlers/disconnectTunnel.ts'
+import { init } from '#src/server/handlers/init.ts'
+import { toprf } from '#src/server/handlers/toprf.ts'
+import type { RPCHandler, RPCType } from '#src/types/index.ts'
 
 export const HANDLERS: { [T in RPCType]: RPCHandler<T> } = {
 	createTunnel,
