--- src/tests/test.tee-bundle.ts
@@ -1,549 +1,0 @@
-/**
- * Test with the new verification bundle that contains attestation docs
- */
-
-import { describe, expect, it } from '@jest/globals'
-import { readFileSync } from 'fs'
-import { join } from 'path'
-import { KOutputPayload, TOutputPayload, VerificationBundle } from 'src/proto/tee-bundle'
-import { reconstructTlsTranscript } from 'src/server/utils/tee-transcript-reconstruction'
-import { logger as LOGGER } from 'src/utils/logger'
-import { makeDefaultOPRFOperator } from 'src/utils/zk'
-
-// Mock logger for testing
-const mockLogger = {
-	info: jest.fn(),
-	debug: jest.fn(),
-	warn: jest.fn(),
-	error: jest.fn()
-}
-
-describe('TEE Bundle Analysis', () => {
-	let standaloneBundleBytes: Uint8Array
-	let teeBundleBytes: Uint8Array
-
-	beforeAll(() => {
-		// Load the standalone verification bundle (development mode with public keys)
-		const standalonePath = join(__dirname, 'verification_bundle.pb')
-		try {
-			standaloneBundleBytes = new Uint8Array(readFileSync(standalonePath))
-			console.log(`Loaded standalone bundle from ${standalonePath}, size: ${standaloneBundleBytes.length} bytes`)
-		} catch(error) {
-			console.error('Failed to load standalone bundle:', error)
-			throw error
-		}
-
-		// Try to load the TEE verification bundle (production mode with attestations)
-		try {
-			const teePath = join(__dirname, 'verification_bundle_tee.pb')
-			teeBundleBytes = new Uint8Array(readFileSync(teePath))
-		} catch(error) {
-			console.warn('TEE attestation bundle not found, will skip TEE tests')
-		}
-	})
-
-	it('should validate standalone bundle structure and OPRF data', async() => {
-		const bundle = VerificationBundle.decode(standaloneBundleBytes)
-
-		// Print bundle structure
-		console.log('\n========== VERIFICATION BUNDLE ANALYSIS ==========')
-		console.log(`Bundle size: ${standaloneBundleBytes.length} bytes`)
-
-		// Validate bundle components
-		expect(bundle).toBeDefined()
-		expect(bundle.teekSigned).toBeDefined()
-		expect(bundle.teekSigned?.body).toBeDefined()
-		expect(bundle.teekSigned?.body?.length).toBeGreaterThan(0)
-
-		// Check bundle mode
-		const hasAttestations = (bundle.teekSigned?.attestationReport?.report && bundle.teekSigned.attestationReport.report.length > 0) ||
-			(bundle.teetSigned?.attestationReport?.report && bundle.teetSigned.attestationReport.report.length > 0)
-		const hasPublicKeys = (bundle.teekSigned?.ethAddress && bundle.teekSigned.ethAddress.length > 0) ||
-			(bundle.teetSigned?.ethAddress && bundle.teetSigned.ethAddress.length > 0)
-
-		console.log('\n--- Bundle Mode ---')
-		console.log(`Has attestations: ${hasAttestations}`)
-		console.log(`Has public keys: ${hasPublicKeys}`)
-		console.log(`Mode: ${hasAttestations ? 'TEE/Production' : 'Standalone/Development'}`)
-
-		// For standalone bundle, should have public keys but not attestations
-		expect(hasPublicKeys).toBe(true)
-		expect(hasAttestations).toBeFalsy()
-
-		// Validate ETH addresses are present
-		expect(bundle.teekSigned?.ethAddress).toBeDefined()
-		expect(bundle.teekSigned?.ethAddress?.length).toBeGreaterThan(0)
-
-		if(bundle.teekSigned?.ethAddress) {
-			const ethAddressHex = '0x' + Buffer.from(bundle.teekSigned.ethAddress).toString('hex')
-			console.log(`TEE_K ETH address: ${ethAddressHex}`)
-		}
-
-		if(bundle.teetSigned?.ethAddress) {
-			const ethAddressHex = '0x' + Buffer.from(bundle.teetSigned.ethAddress).toString('hex')
-			console.log(`TEE_T ETH address: ${ethAddressHex}`)
-		}
-
-		// Analyze the K payload
-		if(bundle.teekSigned?.body) {
-			const kPayload = KOutputPayload.decode(bundle.teekSigned.body)
-
-			console.log('\n--- TEE_K Payload ---')
-			console.log(`Redacted request size: ${kPayload.redactedRequest.length} bytes`)
-			console.log(`Response keystream size: ${kPayload.consolidatedResponseKeystream?.length} bytes`)
-			console.log(`Timestamp: ${new Date(Number(kPayload.timestampMs)).toISOString()}`)
-
-			// Validate K payload structure
-			expect(kPayload.redactedRequest).toBeDefined()
-			expect(kPayload.redactedRequest.length).toBeGreaterThan(0)
-			expect(kPayload.requestRedactionRanges).toBeDefined()
-			expect(Array.isArray(kPayload.requestRedactionRanges)).toBe(true)
-			expect(kPayload.consolidatedResponseKeystream).toBeDefined()
-			expect(kPayload.consolidatedResponseKeystream?.length).toBeGreaterThan(0)
-			expect(kPayload.certificateInfo).toBeDefined()
-			expect(kPayload.timestampMs).toBeDefined()
-			expect(kPayload.timestampMs).toBeGreaterThan(0)
-
-			// Print certificate info
-			if(kPayload.certificateInfo) {
-				console.log('\n--- Certificate Info ---')
-				console.log(`Common Name: ${kPayload.certificateInfo.commonName}`)
-				console.log(`Issuer: ${kPayload.certificateInfo.issuerCommonName}`)
-				console.log(`DNS Names: ${kPayload.certificateInfo.dnsNames.join(', ')}`)
-				console.log(`Valid from: ${new Date(Number(kPayload.certificateInfo.notBeforeUnix) * 1000).toISOString()}`)
-				console.log(`Valid until: ${new Date(Number(kPayload.certificateInfo.notAfterUnix) * 1000).toISOString()}`)
-
-				expect(kPayload.certificateInfo.commonName).toBeDefined()
-				expect(typeof kPayload.certificateInfo.commonName).toBe('string')
-				expect(kPayload.certificateInfo.issuerCommonName).toBeDefined()
-				expect(typeof kPayload.certificateInfo.issuerCommonName).toBe('string')
-				expect(Array.isArray(kPayload.certificateInfo.dnsNames)).toBe(true)
-			}
-
-			// Print redaction ranges
-			console.log('\n--- Request Redaction Ranges ---')
-			console.log(`Total ranges: ${kPayload.requestRedactionRanges.length}`)
-			for(const [idx, range] of kPayload.requestRedactionRanges.entries()) {
-				console.log(`  Range ${idx}: start=${range.start}, length=${range.length}, type=${range.type}`)
-				expect(range.start).toBeGreaterThanOrEqual(0)
-				expect(range.length).toBeGreaterThan(0)
-				expect(typeof range.type).toBe('string')
-				expect(range.type.length).toBeGreaterThan(0)
-			}
-
-			// Print response redaction ranges
-			if(kPayload.responseRedactionRanges && kPayload.responseRedactionRanges.length > 0) {
-				console.log('\n--- Response Redaction Ranges ---')
-				console.log(`Total ranges: ${kPayload.responseRedactionRanges.length}`)
-				for(const [idx, range] of kPayload.responseRedactionRanges.entries()) {
-					console.log(`  Range ${idx}: start=${range.start}, length=${range.length}`)
-				}
-			}
-		}
-
-		// Analyze the T payload (may be missing for standalone mode)
-		if(bundle.teetSigned?.body && bundle.teetSigned.body.length > 0) {
-			const tPayload = TOutputPayload.decode(bundle.teetSigned.body)
-
-			console.log('\n--- TEE_T Payload ---')
-			console.log(`Response ciphertext size: ${tPayload.consolidatedResponseCiphertext?.length} bytes`)
-			console.log(`Timestamp: ${new Date(Number(tPayload.timestampMs)).toISOString()}`)
-			console.log(`Request proof streams: ${tPayload.requestProofStreams?.length || 0}`)
-
-			// Validate T payload structure
-			expect(tPayload.consolidatedResponseCiphertext).toBeDefined()
-			expect(tPayload.consolidatedResponseCiphertext?.length).toBeGreaterThan(0)
-			expect(tPayload.timestampMs).toBeDefined()
-			expect(tPayload.timestampMs).toBeGreaterThan(0)
-		}
-
-		// NEW: Analyze OPRF verification data
-		console.log('\n--- OPRF Verification Data ---')
-		if(bundle.oprfVerifications && bundle.oprfVerifications.length > 0) {
-			console.log(`Total OPRF verifications: ${bundle.oprfVerifications.length}`)
-
-			for(const [idx, oprfData] of bundle.oprfVerifications.entries()) {
-				console.log(`\nOPRF Verification ${idx}:`)
-				console.log(`  Stream position: ${oprfData.streamPos}`)
-				console.log(`  Stream length: ${oprfData.streamLength}`)
-				console.log(`  Public signals size: ${oprfData.publicSignalsJson?.length || 0} bytes`)
-
-				// Try to parse and display public signals if they exist
-				if(oprfData.publicSignalsJson && oprfData.publicSignalsJson.length > 0) {
-					try {
-						const publicSignals = JSON.parse(new TextDecoder().decode(oprfData.publicSignalsJson))
-						console.log('  Public signals (parsed):')
-						console.log(`    ${JSON.stringify(publicSignals, null, 4).split('\n').join('\n    ')}`)
-					} catch(e) {
-						console.log(`  Public signals: [Binary data, ${oprfData.publicSignalsJson.length} bytes]`)
-					}
-				}
-
-				// Validate OPRF data structure
-				expect(oprfData.streamPos).toBeGreaterThanOrEqual(0)
-				expect(oprfData.streamLength).toBeGreaterThan(0)
-			}
-
-			// NEW: Perform actual OPRF verification
-			console.log('\n--- OPRF Verification Test ---')
-			await performOprfVerification(bundle, mockLogger)
-
-		} else {
-			console.log('No OPRF verification data present in bundle')
-		}
-
-		console.log('\n========== END BUNDLE ANALYSIS ==========\n')
-	})
-
-	/**
-	 * Performs actual OPRF verification using the bundle data
-	 */
-	async function performOprfVerification(bundle: VerificationBundle, logger: any) {
-		if(!bundle.oprfVerifications || bundle.oprfVerifications.length === 0) {
-			console.log('No OPRF verifications to test')
-			return
-		}
-
-		// Get TEE payloads for ciphertext extraction
-		const tPayload = bundle.teetSigned?.body ? TOutputPayload.decode(bundle.teetSigned.body) : null
-		if(!tPayload?.consolidatedResponseCiphertext) {
-			console.log('Missing TEE_T payload for ciphertext extraction')
-			return
-		}
-
-		for(const [idx, oprfData] of bundle.oprfVerifications.entries()) {
-			console.log(`\nVerifying OPRF ${idx}...`)
-
-			try {
-				// Parse the public signals
-				const publicSignalsJson = JSON.parse(new TextDecoder().decode(oprfData.publicSignalsJson))
-				const { proof, publicSignals } = publicSignalsJson
-
-				// Debug: Check what fields are present and their values
-				console.log('Available fields in publicSignals:', Object.keys(publicSignals))
-				console.log('Blocks[0] nonce:', publicSignals.blocks?.[0]?.nonce)
-				console.log('TOPRF domainSeparator:', publicSignals.toprf?.domainSeparator)
-				console.log('TOPRF output:', publicSignals.toprf?.output)
-				console.log('First response publicKeyShare:', publicSignals.toprf?.responses?.[0]?.publicKeyShare)
-				console.log('TOPRF mask:', publicSignals.toprf?.mask)
-
-				// Extract ciphertext chunk from TEE_T payload
-				const ciphertextChunk = tPayload.consolidatedResponseCiphertext.slice(
-					oprfData.streamPos,
-					oprfData.streamPos + oprfData.streamLength
-				)
-				console.log(`Extracted ${ciphertextChunk.length} bytes from stream position ${oprfData.streamPos}`)
-
-				// Build complete public signals for verification (excluding private mask/key)
-				// Replace the null input field with extracted ciphertext
-				const completePublicSignals = {
-					...publicSignals,
-					input: ciphertextChunk, // Replace null input with extracted ciphertext
-					blocks: publicSignals.blocks.map((block: any) => {
-						if(!block.nonce) {
-							throw new Error('Block nonce is undefined')
-						}
-
-						return {
-							nonce: Buffer.from(block.nonce, 'base64'),
-							counter: block.counter
-						}
-					}),
-					toprf: {
-						...publicSignals.toprf,
-						domainSeparator: publicSignals.toprf.domainSeparator ?
-							Buffer.from(publicSignals.toprf.domainSeparator, 'base64').toString('utf8') :
-							'reclaim',
-						output: publicSignals.toprf.output ?
-							Buffer.from(publicSignals.toprf.output, 'base64') :
-							new Uint8Array(),
-						responses: publicSignals.toprf.responses.map((resp: any) => {
-							if(!resp.publicKeyShare) {
-								throw new Error('Response publicKeyShare is undefined')
-							}
-
-							if(!resp.evaluated) {
-								throw new Error('Response evaluated is undefined')
-							}
-
-							if(!resp.c) {
-								throw new Error('Response c is undefined')
-							}
-
-							if(!resp.r) {
-								throw new Error('Response r is undefined')
-							}
-
-							return {
-								publicKeyShare: Buffer.from(resp.publicKeyShare, 'base64'),
-								evaluated: Buffer.from(resp.evaluated, 'base64'),
-								c: Buffer.from(resp.c, 'base64'),
-								r: Buffer.from(resp.r, 'base64')
-							}
-						})
-						// mask is private and omitted from verification
-					}
-				}
-
-				console.log(`Domain separator: "${completePublicSignals.toprf.domainSeparator}"`)
-				console.log(`OPRF location: pos=${completePublicSignals.toprf.locations[0].pos}, len=${completePublicSignals.toprf.locations[0].len}`)
-
-				// Get OPRF operator for the cipher type
-				const algorithm = publicSignals.cipher?.replace('-toprf', '') || 'aes-256-ctr'
-				console.log(`Using algorithm: ${algorithm}`)
-
-				const oprfOperator = makeDefaultOPRFOperator(algorithm, 'gnark', LOGGER)
-
-				// Convert proof from base64
-				const proofBytes = Buffer.from(proof, 'base64')
-
-				// Debug: Check final structure before verification
-				console.log('Final input length:', completePublicSignals.input?.length)
-				console.log('Proof length:', proofBytes?.length)
-				console.log('First nonce type:', typeof completePublicSignals.blocks[0].nonce)
-				console.log('Domain separator type:', typeof completePublicSignals.toprf.domainSeparator)
-
-				// Perform verification
-				console.log('About to call groth16Verify...')
-				console.log('Operator type:', typeof oprfOperator.groth16Verify)
-
-				// Try verification with error handling for each step
-				const isValid = await oprfOperator.groth16Verify(
-					completePublicSignals,
-					proofBytes,
-					logger
-				)
-
-				console.log(`OPRF verification result: ${isValid ? 'VALID ✓' : 'INVALID ✗'}`)
-				// Don't expect true for now, just check it doesn't crash
-				console.log('Verification completed without crash!')
-
-			} catch(error) {
-				console.error(`OPRF verification failed for index ${idx}:`, error)
-				// Don't fail the test, just log the error for now
-				console.log(`OPRF verification skipped due to error: ${error.message}`)
-			}
-		}
-	}
-
-	it('should reconstruct TLS transcript from standalone bundle', async() => {
-		const bundle = VerificationBundle.decode(standaloneBundleBytes)
-
-		// Check if we have both TEE_K and TEE_T payloads
-		if(!bundle.teekSigned?.body || bundle.teekSigned.body.length === 0) {
-			// Skip if missing K payload
-			return
-		}
-
-		if(!bundle.teetSigned?.body || bundle.teetSigned.body.length === 0) {
-			// Skip if missing T payload (expected for standalone bundle)
-			return
-		}
-
-		const kOutputPayload = KOutputPayload.decode(bundle.teekSigned.body)
-		const tOutputPayload = TOutputPayload.decode(bundle.teetSigned.body)
-
-		// Add timestamps to payloads if they don't exist (for older test bundles)
-		const now = Date.now()
-		if(!kOutputPayload.timestampMs) {
-			kOutputPayload.timestampMs = now
-		}
-
-		if(!tOutputPayload.timestampMs) {
-			tOutputPayload.timestampMs = now
-		}
-
-		// Create mock bundle data
-		const mockTeeBundleData = {
-			teekSigned: bundle.teekSigned,
-			teetSigned: bundle.teetSigned,
-			teekPublicKey: new Uint8Array(64), // Mock public key
-			teetPublicKey: new Uint8Array(64), // Mock public key
-			kOutputPayload,
-			tOutputPayload,
-			handshakeKeys: undefined, // handshakeKeys removed from new bundle format
-		}
-
-		const transcriptData = await reconstructTlsTranscript(mockTeeBundleData as any, mockLogger as any)
-
-		// Validate transcript reconstruction results
-		expect(transcriptData).toBeDefined()
-		expect(transcriptData.revealedRequest).toBeDefined()
-		expect(transcriptData.reconstructedResponse).toBeDefined()
-		expect(transcriptData.revealedRequest.length).toBeGreaterThan(0)
-		expect(transcriptData.reconstructedResponse.length).toBeGreaterThan(0)
-		expect(transcriptData.certificateInfo).toBeDefined()
-
-		// Validate revealed request is valid HTTP
-		const requestText = new TextDecoder('utf-8', { fatal: false }).decode(transcriptData.revealedRequest)
-		expect(requestText.length).toBeGreaterThan(0)
-
-		// Look for HTTP details and validate structure
-		const httpMatch = requestText.match(/^(GET|POST|PUT|DELETE|PATCH)\s+(\S+)\s+HTTP\/[\d.]+/m)
-		const hostMatch = requestText.match(/Host:\s*([^\r\n]+)/i)
-
-		if(httpMatch) {
-			expect(httpMatch[1]).toMatch(/^(GET|POST|PUT|DELETE|PATCH)$/)
-			expect(httpMatch[2]).toMatch(/^\//)
-		}
-
-		if(hostMatch) {
-			expect(hostMatch[1].trim().length).toBeGreaterThan(0)
-		}
-
-		// Validate response is valid HTTP
-		const responseText = new TextDecoder('utf-8', { fatal: false }).decode(transcriptData.reconstructedResponse)
-		expect(responseText.length).toBeGreaterThan(0)
-
-		const statusMatch = responseText.match(/^HTTP\/[\d.]+\s+(\d+)\s+([^\r\n]+)/m)
-		if(statusMatch) {
-			const statusCode = parseInt(statusMatch[1])
-			expect(statusCode).toBeGreaterThanOrEqual(100)
-			expect(statusCode).toBeLessThan(600)
-			expect(statusMatch[2].trim().length).toBeGreaterThan(0)
-		}
-
-		// Validate certificate info structure
-		if(transcriptData.certificateInfo) {
-			expect(transcriptData.certificateInfo.commonName).toBeDefined()
-			expect(typeof transcriptData.certificateInfo.commonName).toBe('string')
-		}
-	})
-
-	it('should validate TEE attestation bundle structure', () => {
-		if(!teeBundleBytes) {
-			// Skip if bundle not available
-			return
-		}
-
-		const bundle = VerificationBundle.decode(teeBundleBytes)
-
-		// Validate bundle components
-		expect(bundle).toBeDefined()
-		expect(bundle.teekSigned).toBeDefined()
-		expect(bundle.teekSigned?.body).toBeDefined()
-		expect(bundle.teekSigned?.body?.length).toBeGreaterThan(0)
-
-		// Check bundle mode
-		const hasAttestations = (bundle.teekSigned?.attestationReport?.report && bundle.teekSigned.attestationReport.report.length > 0) ||
-			(bundle.teetSigned?.attestationReport?.report && bundle.teetSigned.attestationReport.report.length > 0)
-		const hasPublicKeys = (bundle.teekSigned?.ethAddress && bundle.teekSigned.ethAddress.length > 0) ||
-			(bundle.teetSigned?.ethAddress && bundle.teetSigned.ethAddress.length > 0)
-
-		// For TEE bundle, should have either attestations or public keys (or both)
-		expect(hasAttestations || hasPublicKeys).toBe(true)
-
-		if(hasAttestations) {
-			// Validate attestation structure
-			if(bundle.teekSigned?.attestationReport) {
-				expect(bundle.teekSigned.attestationReport.report).toBeDefined()
-				expect(bundle.teekSigned.attestationReport.report.length).toBeGreaterThan(0)
-				expect(bundle.teekSigned.attestationReport.type).toBeDefined()
-				expect(typeof bundle.teekSigned.attestationReport.type).toBe('string')
-			}
-
-			if(bundle.teetSigned?.attestationReport) {
-				expect(bundle.teetSigned.attestationReport.report).toBeDefined()
-				expect(bundle.teetSigned.attestationReport.report.length).toBeGreaterThan(0)
-				expect(bundle.teetSigned.attestationReport.type).toBeDefined()
-				expect(typeof bundle.teetSigned.attestationReport.type).toBe('string')
-			}
-		}
-
-		if(hasPublicKeys) {
-			// Validate public key structure
-			if(bundle.teekSigned?.ethAddress) {
-				expect(bundle.teekSigned.ethAddress.length).toBeGreaterThan(0)
-			}
-
-			if(bundle.teetSigned?.ethAddress) {
-				expect(bundle.teetSigned.ethAddress.length).toBeGreaterThan(0)
-			}
-		}
-
-		// Validate K payload
-		if(bundle.teekSigned?.body) {
-			const kPayload = KOutputPayload.decode(bundle.teekSigned.body)
-
-			// Validate K payload structure
-			expect(kPayload.redactedRequest).toBeDefined()
-			expect(kPayload.redactedRequest.length).toBeGreaterThan(0)
-			expect(kPayload.requestRedactionRanges).toBeDefined()
-			expect(Array.isArray(kPayload.requestRedactionRanges)).toBe(true)
-			expect(kPayload.consolidatedResponseKeystream).toBeDefined()
-			expect(kPayload.consolidatedResponseKeystream?.length).toBeGreaterThan(0)
-			expect(kPayload.certificateInfo).toBeDefined()
-			expect(kPayload.timestampMs).toBeDefined()
-			expect(kPayload.timestampMs).toBeGreaterThan(0)
-
-			// Validate certificate info
-			if(kPayload.certificateInfo) {
-				expect(kPayload.certificateInfo.commonName).toBeDefined()
-				expect(typeof kPayload.certificateInfo.commonName).toBe('string')
-				expect(kPayload.certificateInfo.issuerCommonName).toBeDefined()
-				expect(typeof kPayload.certificateInfo.issuerCommonName).toBe('string')
-				expect(Array.isArray(kPayload.certificateInfo.dnsNames)).toBe(true)
-			}
-
-			// Validate redaction ranges
-			for(const range of kPayload.requestRedactionRanges) {
-				expect(range.start).toBeGreaterThanOrEqual(0)
-				expect(range.length).toBeGreaterThan(0)
-				expect(typeof range.type).toBe('string')
-				expect(range.type.length).toBeGreaterThan(0)
-			}
-		}
-
-		// Validate T payload
-		if(bundle.teetSigned?.body && bundle.teetSigned.body.length > 0) {
-			const tPayload = TOutputPayload.decode(bundle.teetSigned.body)
-
-			// Validate T payload structure
-			expect(tPayload.consolidatedResponseCiphertext).toBeDefined()
-			expect(tPayload.consolidatedResponseCiphertext?.length).toBeGreaterThan(0)
-			expect(tPayload.timestampMs).toBeDefined()
-			expect(tPayload.timestampMs).toBeGreaterThan(0)
-		}
-	})
-
-	it('should reconstruct TLS transcript from TEE attestation bundle', async() => {
-		if(!teeBundleBytes) {
-			// Skip if bundle not available
-			return
-		}
-
-		const bundle = VerificationBundle.decode(teeBundleBytes)
-		const kOutputPayload = KOutputPayload.decode(bundle.teekSigned!.body)
-		const tOutputPayload = TOutputPayload.decode(bundle.teetSigned!.body)
-
-		// Add timestamps to payloads if they don't exist (for older test bundles)
-		const now = Date.now()
-		if(!kOutputPayload.timestampMs) {
-			kOutputPayload.timestampMs = now
-		}
-
-		if(!tOutputPayload.timestampMs) {
-			tOutputPayload.timestampMs = now
-		}
-
-		// Create mock bundle data
-		const mockTeeBundleData = {
-			teekSigned: bundle.teekSigned!,
-			teetSigned: bundle.teetSigned!,
-			teekPublicKey: new Uint8Array(64), // Mock public key
-			teetPublicKey: new Uint8Array(64), // Mock public key
-			kOutputPayload,
-			tOutputPayload,
-			handshakeKeys: undefined, // handshakeKeys removed from new bundle format
-		}
-
-		const transcriptData = await reconstructTlsTranscript(mockTeeBundleData as any, mockLogger as any)
-
-		// Validate transcript reconstruction results
-		expect(transcriptData).toBeDefined()
-		expect(transcriptData.revealedRequest).toBeDefined()
-		expect(transcriptData.reconstructedResponse).toBeDefined()
-		expect(transcriptData.revealedRequest.length).toBeGreaterThan(0)
-		expect(transcriptData.reconstructedResponse.length).toBeGreaterThan(0)
-	})
-})
