--- src/tests/test.tee-signatures.ts
@@ -1,247 +1,0 @@
-/**
- * TEE Signature Verification Test
- * Tests the complete TEE bundle verification including ETH signature validation
- */
-
-import { describe, expect, it } from '@jest/globals'
-import { readFileSync } from 'fs'
-import { join } from 'path'
-import { VerificationBundle } from 'src/proto/tee-bundle'
-import { verifyTeeBundle } from 'src/server/utils/tee-verification'
-
-// Mock logger for testing
-const mockLogger = {
-	info: (msg: any) => console.log('[INFO]', msg),
-	debug: (msg: any) => console.log('[DEBUG]', msg),
-	warn: (msg: any) => console.log('[WARN]', msg),
-	error: (msg: any) => console.log('[ERROR]', msg)
-}
-
-describe('TEE Signature Verification', () => {
-	let standaloneBundleBytes: Uint8Array
-	let teeBundleBytes: Uint8Array
-
-	beforeAll(() => {
-		// Load the standalone verification bundle (development mode with public keys)
-		const standalonePath = join(__dirname, 'verification_bundle.pb')
-		standaloneBundleBytes = new Uint8Array(readFileSync(standalonePath))
-
-		// Try to load the TEE verification bundle (production mode with attestations)
-		try {
-			const teePath = join(__dirname, 'verification_bundle_tee.pb')
-			teeBundleBytes = new Uint8Array(readFileSync(teePath))
-		} catch(error) {
-			console.warn('TEE attestation bundle not found, skipping TEE attestation tests')
-		}
-	})
-
-	it('should validate standalone mode bundles with embedded public keys', async() => {
-		const bundle = VerificationBundle.decode(standaloneBundleBytes)
-
-		// Validate bundle structure
-		expect(bundle).toBeDefined()
-		expect(bundle.teekSigned).toBeDefined()
-		expect(bundle.teetSigned).toBeDefined()
-
-		// Check bundle mode
-		const hasAttestations = (bundle.teekSigned?.attestationReport?.report && bundle.teekSigned.attestationReport.report.length > 0) ||
-			(bundle.teetSigned?.attestationReport?.report && bundle.teetSigned.attestationReport.report.length > 0)
-
-		const hasEthAddresses = (bundle.teekSigned?.ethAddress && bundle.teekSigned.ethAddress.length > 0) ||
-			(bundle.teetSigned?.ethAddress && bundle.teetSigned.ethAddress.length > 0)
-
-		// For standalone mode, should have ETH addresses but not attestations
-		expect(hasEthAddresses).toBe(true)
-		expect(hasAttestations).toBeFalsy()
-
-		// Validate TEE_K ETH address
-		if(bundle.teekSigned?.ethAddress) {
-			expect(bundle.teekSigned.ethAddress.length).toBeGreaterThan(0)
-			// ETH address can be either 20 bytes (raw ETH address) or 42 bytes (hex-encoded "0x..." string)
-			expect([20, 42]).toContain(bundle.teekSigned.ethAddress.length)
-
-			if(bundle.teekSigned.ethAddress.length === 42) {
-				// Should be hex-encoded ETH address
-				const addressStr = new TextDecoder().decode(bundle.teekSigned.ethAddress)
-				expect(addressStr).toMatch(/^0x[0-9a-fA-F]{40}$/)
-			}
-		}
-
-		// Validate TEE_T ETH address
-		if(bundle.teetSigned?.ethAddress) {
-			expect(bundle.teetSigned.ethAddress.length).toBeGreaterThan(0)
-			// ETH address can be either 20 bytes (raw ETH address) or 42 bytes (hex-encoded "0x..." string)
-			expect([20, 42]).toContain(bundle.teetSigned.ethAddress.length)
-
-			if(bundle.teetSigned.ethAddress.length === 42) {
-				// Should be hex-encoded ETH address
-				const addressStr = new TextDecoder().decode(bundle.teetSigned.ethAddress)
-				expect(addressStr).toMatch(/^0x[0-9a-fA-F]{40}$/)
-			}
-		}
-	})
-
-	it('should verify TEE bundle signatures using embedded public keys', async() => {
-		// Mock Date.now() to return a time close to the bundle timestamps for validation
-		const originalDateNow = Date.now
-		const mockTime = 1755774883738 + (2 * 60 * 1000) // 2 minutes after TEE_K timestamp
-		Date.now = jest.fn(() => mockTime)
-
-		try {
-			// This should work with embedded public keys (standalone mode)
-			const result = await verifyTeeBundle(standaloneBundleBytes, mockLogger as any)
-
-			// Validate verification results
-			expect(result).toBeDefined()
-			expect(result.teekSigned).toBeDefined()
-			expect(result.teetSigned).toBeDefined()
-			expect(result.kOutputPayload).toBeDefined()
-			expect(result.tOutputPayload).toBeDefined()
-
-			// Validate that signatures are present and valid (non-empty)
-			expect(result.teekSigned.signature).toBeDefined()
-			expect(result.teekSigned.signature.length).toBeGreaterThan(0)
-			expect(result.teetSigned.signature).toBeDefined()
-			expect(result.teetSigned.signature.length).toBeGreaterThan(0)
-
-			// Validate payload contents
-			expect(result.kOutputPayload.redactedRequest).toBeDefined()
-			expect(result.kOutputPayload.redactedRequest.length).toBeGreaterThan(0)
-			expect(result.kOutputPayload.timestampMs).toBeGreaterThan(0)
-
-			expect(result.tOutputPayload.timestampMs).toBeGreaterThan(0)
-
-		} finally {
-			// Restore original Date.now()
-			Date.now = originalDateNow
-		}
-	})
-
-	it('should validate signature structure and metadata', () => {
-		const bundle = VerificationBundle.decode(standaloneBundleBytes)
-
-		// Validate TEE_K signed data
-		if(bundle.teekSigned) {
-			expect(bundle.teekSigned.body).toBeDefined()
-			expect(bundle.teekSigned.body.length).toBeGreaterThan(0)
-			expect(bundle.teekSigned.signature).toBeDefined()
-			expect(bundle.teekSigned.signature.length).toBeGreaterThan(0)
-			expect(bundle.teekSigned.bodyType).toBeDefined()
-			expect(typeof bundle.teekSigned.bodyType).toBe('number')
-			expect(bundle.teekSigned.bodyType).toBeGreaterThanOrEqual(0)
-
-			// Validate signature format (should be 64 bytes for ECDSA)
-			expect(bundle.teekSigned.signature.length).toBeGreaterThanOrEqual(64)
-		}
-
-		// Validate TEE_T signed data
-		if(bundle.teetSigned) {
-			expect(bundle.teetSigned.body).toBeDefined()
-			expect(bundle.teetSigned.body.length).toBeGreaterThan(0)
-			expect(bundle.teetSigned.signature).toBeDefined()
-			expect(bundle.teetSigned.signature.length).toBeGreaterThan(0)
-			expect(bundle.teetSigned.bodyType).toBeDefined()
-			expect(typeof bundle.teetSigned.bodyType).toBe('number')
-			expect(bundle.teetSigned.bodyType).toBeGreaterThanOrEqual(0)
-
-			// Validate signature format (should be 64 bytes for ECDSA)
-			expect(bundle.teetSigned.signature.length).toBeGreaterThanOrEqual(64)
-		}
-	})
-
-	it('should validate TEE attestation bundles structure', async() => {
-		if(!teeBundleBytes) {
-			// Skip if bundle not available
-			return
-		}
-
-		const bundle = VerificationBundle.decode(teeBundleBytes)
-
-		// Validate bundle structure
-		expect(bundle).toBeDefined()
-		expect(bundle.teekSigned).toBeDefined()
-		expect(bundle.teetSigned).toBeDefined()
-
-		// Check bundle mode
-		const hasAttestations = (bundle.teekSigned?.attestationReport?.report && bundle.teekSigned.attestationReport.report.length > 0) ||
-			(bundle.teetSigned?.attestationReport?.report && bundle.teetSigned.attestationReport.report.length > 0)
-
-		const hasEthAddresses = (bundle.teekSigned?.ethAddress && bundle.teekSigned.ethAddress.length > 0) ||
-			(bundle.teetSigned?.ethAddress && bundle.teetSigned.ethAddress.length > 0)
-
-		// Bundle must have either attestations or ETH addresses
-		expect(hasAttestations || hasEthAddresses).toBe(true)
-
-		if(hasAttestations) {
-			// Validate attestation structure
-			if(bundle.teekSigned?.attestationReport) {
-				expect(bundle.teekSigned.attestationReport.type).toBeDefined()
-				expect(typeof bundle.teekSigned.attestationReport.type).toBe('string')
-				expect(bundle.teekSigned.attestationReport.type.length).toBeGreaterThan(0)
-				expect(bundle.teekSigned.attestationReport.report).toBeDefined()
-				expect(bundle.teekSigned.attestationReport.report.length).toBeGreaterThan(0)
-			}
-
-			if(bundle.teetSigned?.attestationReport) {
-				expect(bundle.teetSigned.attestationReport.type).toBeDefined()
-				expect(typeof bundle.teetSigned.attestationReport.type).toBe('string')
-				expect(bundle.teetSigned.attestationReport.type.length).toBeGreaterThan(0)
-				expect(bundle.teetSigned.attestationReport.report).toBeDefined()
-				expect(bundle.teetSigned.attestationReport.report.length).toBeGreaterThan(0)
-			}
-		}
-
-		if(hasEthAddresses) {
-			// Validate ETH address structure
-			if(bundle.teekSigned?.ethAddress) {
-				expect(bundle.teekSigned.ethAddress.length).toBeGreaterThan(0)
-			}
-
-			if(bundle.teetSigned?.ethAddress) {
-				expect(bundle.teetSigned.ethAddress.length).toBeGreaterThan(0)
-			}
-		}
-	})
-
-	it('should verify TEE bundle with attestations or handle expected test failures', async() => {
-
-		// Mock Date.now() to return a time close to the bundle timestamps for validation
-		const originalDateNow = Date.now
-		const mockTime = 1755698083175 + (2 * 60 * 1000) // 2 minutes after TEE_K timestamp in the TEE bundle
-		Date.now = jest.fn(() => mockTime)
-
-		try {
-			// This should work with TEE attestations (production mode)
-			const result = await verifyTeeBundle(teeBundleBytes, mockLogger as any)
-
-			// Validate verification results if successful
-			expect(result).toBeDefined()
-			expect(result.teekSigned).toBeDefined()
-			expect(result.teetSigned).toBeDefined()
-			expect(result.kOutputPayload).toBeDefined()
-			expect(result.tOutputPayload).toBeDefined()
-
-			// Validate payload structure
-			expect(result.kOutputPayload.redactedRequest).toBeDefined()
-			expect(result.kOutputPayload.redactedRequest.length).toBeGreaterThan(0)
-			expect(result.kOutputPayload.timestampMs).toBeGreaterThan(0)
-			expect(result.tOutputPayload.timestampMs).toBeGreaterThan(0)
-
-		} catch(error) {
-			// Check if it's expected errors with test data
-			if((error as Error).message.includes('attestation')) {
-				// For attestation errors with test data, this is expected behavior
-				// The test bundle likely has demo/test attestations that won't pass real validation
-				expect(error).toBeInstanceOf(Error)
-				expect((error as Error).message).toContain('attestation')
-				return // This is expected for test bundles
-			}
-
-			// For other errors, re-throw
-			throw error
-		} finally {
-			// Restore original Date.now()
-			Date.now = originalDateNow
-		}
-	})
-})
